+++ src/Mailer.php
@@ -112,9 +112,9 @@ abstract class Mailer implements MailerInterface
      *
      * @param MessageInterface[] $messages List of email messages, which should be sent.
      *
-     * @return MultipleSendResult The result object that contains all messages and errors for failed sent messages.
+     * @return SendResults The result object that contains all messages and errors for failed sent messages.
      */
-    public function sendMultiple(array $messages): MultipleSendResult
+    public function sendMultiple(array $messages): SendResults
     {
         $successMessages = [];
         $failMessages = [];
@@ -129,7 +129,7 @@ abstract class Mailer implements MailerInterface
             $successMessages[] = $message;
         }
 
-        return new MultipleSendResult($successMessages, $failMessages);
+        return new SendResults($successMessages, $failMessages);
     }
 
     /**